refactoring: move some code to better locations
